server/ml-api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.26.4
tensorflow==2.18.0
torch== 2.0.1
opencv-python==4.8.1.78
ultralytics==8.3.145
requests==2.31.0